fix(schema): resolve 26 database checker errors and adopt date-based SQL versioning (fixes #344)
- Replace MODIFY COLUMN with MODIFY in SQL updates (Joomla parser misreads COLUMN keyword as column name)
- Split multi-column ALTER TABLE into separate statements (parser only handles one MODIFY per ALTER)
- Fix ordering column type mismatch in 6.0.27.sql and 6.1.2.sql (tinyint→int to match current state)
- Adopt Joomla core date-based schema versioning: 6.1.5-2026-04-04.sql
- Update build_package.php validateSqlAlignment() to handle both old and new naming patterns
- Add repairMissingTables() to install script for broken upgrade detection

+++ administrator/components/com_j2commerce/sql/updates/mysql/6.1.2.sql
@@ -13,7 +13,7 @@ ALTER TABLE `#__j2commerce_product_options`
   MODIFY `parent_id` int NOT NULL DEFAULT 0;
 
 ALTER TABLE `#__j2commerce_product_options`
-  MODIFY `ordering` tinyint NOT NULL DEFAULT 0;
+  MODIFY `ordering` int NOT NULL DEFAULT 0;
 
 ALTER TABLE `#__j2commerce_product_options`
   MODIFY `required` int NOT NULL DEFAULT 0;
